+++ test/ServiceFactoryTest.php
@@ -1,36 +1,35 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-navigation for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-navigation/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-navigation/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Navigation;
+namespace LaminasTest\Navigation;
 
-use Zend\Config\Config;
-use Zend\Http\Request as HttpRequest;
-use Zend\Mvc\Application;
-use Zend\Mvc\MvcEvent;
-use Zend\Mvc\Router\RouteMatch;
-use Zend\Mvc\Router\RouteStackInterface;
-use Zend\Navigation;
-use Zend\Navigation\Page\Mvc as MvcPage;
-use Zend\Navigation\Service\ConstructedNavigationFactory;
-use Zend\Navigation\Service\DefaultNavigationFactory;
-use Zend\Navigation\Service\NavigationAbstractServiceFactory;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\Config\Config;
+use Laminas\Http\Request as HttpRequest;
+use Laminas\Mvc\Application;
+use Laminas\Mvc\MvcEvent;
+use Laminas\Mvc\Router\RouteMatch;
+use Laminas\Mvc\Router\RouteStackInterface;
+use Laminas\Navigation;
+use Laminas\Navigation\Page\Mvc as MvcPage;
+use Laminas\Navigation\Service\ConstructedNavigationFactory;
+use Laminas\Navigation\Service\DefaultNavigationFactory;
+use Laminas\Navigation\Service\NavigationAbstractServiceFactory;
+use Laminas\ServiceManager\ServiceManager;
 
 /**
- * Tests the class Zend\Navigation\MvcNavigationFactory
+ * Tests the class Laminas\Navigation\MvcNavigationFactory
  *
- * @group      Zend_Navigation
+ * @group      Laminas_Navigation
  */
 class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var \Zend\ServiceManager\ServiceManager
+     * @var \Laminas\ServiceManager\ServiceManager
      */
     protected $serviceManager;
 
@@ -41,8 +40,8 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     {
         if (! class_exists(Application::class)) {
             $this->markTestSkipped(
-                'Skipping zend-mvc-related tests until that component is updated '
-                . 'to be forwards-compatible with zend-servicemanager v3'
+                'Skipping laminas-mvc-related tests until that component is updated '
+                . 'to be forwards-compatible with laminas-servicemanager v3'
             );
         }
 
@@ -97,7 +96,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\AbstractNavigationFactory
+     * @covers \Laminas\Navigation\Service\AbstractNavigationFactory
      */
     public function testDefaultFactoryAcceptsFileString()
     {
@@ -106,7 +105,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\DefaultNavigationFactory
+     * @covers \Laminas\Navigation\Service\DefaultNavigationFactory
      */
     public function testMvcPagesGetInjectedWithComponents()
     {
@@ -116,8 +115,8 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
         $recursive = function ($that, $pages) use (&$recursive) {
             foreach ($pages as $page) {
                 if ($page instanceof MvcPage) {
-                    $that->assertInstanceOf('Zend\Mvc\Router\RouteStackInterface', $page->getRouter());
-                    $that->assertInstanceOf('Zend\Mvc\Router\RouteMatch', $page->getRouteMatch());
+                    $that->assertInstanceOf('Laminas\Mvc\Router\RouteStackInterface', $page->getRouter());
+                    $that->assertInstanceOf('Laminas\Mvc\Router\RouteMatch', $page->getRouteMatch());
                 }
 
                 $recursive($that, $page->getPages());
@@ -127,7 +126,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\ConstructedNavigationFactory
+     * @covers \Laminas\Navigation\Service\ConstructedNavigationFactory
      */
     public function testConstructedNavigationFactoryInjectRouterAndMatcher()
     {
@@ -141,8 +140,8 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
                 ->method('injectComponents')
                 ->with(
                     $this->isType('array'),
-                    $this->isInstanceOf('Zend\Mvc\Router\RouteMatch'),
-                    $this->isInstanceOf('Zend\Mvc\Router\RouteStackInterface')
+                    $this->isInstanceOf('Laminas\Mvc\Router\RouteMatch'),
+                    $this->isInstanceOf('Laminas\Mvc\Router\RouteStackInterface')
                 );
 
         $this->serviceManager->setFactory('Navigation', function ($services) use ($factory) {
@@ -153,7 +152,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\ConstructedNavigationFactory
+     * @covers \Laminas\Navigation\Service\ConstructedNavigationFactory
      */
     public function testMvcPagesGetInjectedWithComponentsInConstructedNavigationFactory()
     {
@@ -167,8 +166,8 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
         $recursive = function ($that, $pages) use (&$recursive) {
             foreach ($pages as $page) {
                 if ($page instanceof MvcPage) {
-                    $that->assertInstanceOf('Zend\Mvc\Router\RouteStackInterface', $page->getRouter());
-                    $that->assertInstanceOf('Zend\Mvc\Router\RouteMatch', $page->getRouteMatch());
+                    $that->assertInstanceOf('Laminas\Mvc\Router\RouteStackInterface', $page->getRouter());
+                    $that->assertInstanceOf('Laminas\Mvc\Router\RouteMatch', $page->getRouteMatch());
                 }
 
                 $recursive($that, $page->getPages());
@@ -178,7 +177,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\DefaultNavigationFactory
+     * @covers \Laminas\Navigation\Service\DefaultNavigationFactory
      */
     public function testDefaultFactory()
     {
@@ -189,7 +188,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\ConstructedNavigationFactory
+     * @covers \Laminas\Navigation\Service\ConstructedNavigationFactory
      */
     public function testConstructedFromArray()
     {
@@ -216,7 +215,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\ConstructedNavigationFactory
+     * @covers \Laminas\Navigation\Service\ConstructedNavigationFactory
      */
     public function testConstructedFromFileString()
     {
@@ -229,7 +228,7 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\ConstructedNavigationFactory
+     * @covers \Laminas\Navigation\Service\ConstructedNavigationFactory
      */
     public function testConstructedFromConfig()
     {
@@ -256,25 +255,25 @@ class ServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers \Zend\Navigation\Service\NavigationAbstractServiceFactory
+     * @covers \Laminas\Navigation\Service\NavigationAbstractServiceFactory
      */
     public function testNavigationAbstractServiceFactory()
     {
         $factory = new NavigationAbstractServiceFactory();
 
         $this->assertTrue(
-            $factory->canCreate($this->serviceManager, 'Zend\Navigation\File')
+            $factory->canCreate($this->serviceManager, 'Laminas\Navigation\File')
         );
         $this->assertFalse(
-            $factory->canCreate($this->serviceManager, 'Zend\Navigation\Unknown')
+            $factory->canCreate($this->serviceManager, 'Laminas\Navigation\Unknown')
         );
 
         $container = $factory(
             $this->serviceManager,
-            'Zend\Navigation\File'
+            'Laminas\Navigation\File'
         );
 
-        $this->assertInstanceOf('Zend\Navigation\Navigation', $container);
+        $this->assertInstanceOf('Laminas\Navigation\Navigation', $container);
         $this->assertEquals(3, $container->count());
     }
 }