+++ CHANGELOG.md
@@ -18,6 +18,6 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#5](https://github.com/zendframework/zend-navigation/pull/5) and
-  [#20](https://github.com/zendframework/zend-navigation/pull/20) update the
-  code to be forwards compatible with zend-servicemanager v3.
+- [zendframework/zend-navigation#5](https://github.com/zendframework/zend-navigation/pull/5) and
+  [zendframework/zend-navigation#20](https://github.com/zendframework/zend-navigation/pull/20) update the
+  code to be forwards compatible with laminas-servicemanager v3.